requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
tzdata==2024.1
itsdangerous==2.2.0
qrcode==7.4.2
Pillow==10.4.0



